Make clean_roms.sh buffer output

+++ clean_roms.sh
@@ -8,9 +8,28 @@ shopt -s nullglob
 declare -r MAME_BIN="${MAME_BIN:-"$(which mame)"}"
 declare -r MAME_ROMS_DIR="${MAME_ROMS_DIR:-"/usr/share/games/mame/roms"}"
 
-for rom in $(${MAME_BIN} -verifyroms | grep ' is bad ' | awk ' {print $2 ".zip"} '); do
+declare -r OPTION="${1}"
+
+echo "${script_name}[INFO]: Detecting bad roms..."
+declare -a bad_roms=()
+declare -i detected_roms=0
+while read -r line; do
+    bad_rom=$(echo "${line}" | awk '/is bad/ {print $2".zip"}')
+    if [ -n "${bad_rom}" ]; then
+        echo "${script_name}[INFO]: ${bad_rom} is bad"
+        bad_roms+="${bad_rom}" 
+        ((detected_roms++))
+    fi
+done < <(stdbuf -oL ${MAME_BIN} -verifyroms)
+echo "${script_name}[INFO]: Detected ${detected_roms} bad roms."
+
+if [ "${OPTION,,}" != "delete" ]; then
+    return 0
+fi
+
+declare -i deleted_roms=0
+for rom in "${bad_roms[@]}"; do
     declare rom_path="${MAME_ROMS_DIR}/${rom}"
-    declare -i deleted_roms=0
     if [ -f "${rom_path}" ]; then
         sudo rm "${rom_path}"
         echo "${script_name}[INFO]: Deleted '${rom_path}'"
